ML-4 Ensemble with Banking
The data is related with direct marketing campaigns of a Portuguese banking institution.

Domain : Banking

Leveraging customer information is paramount for most businesses. In the case of a bank, attributes of customers like the ones mentioned below can be crucial in strategizing a marketing campaign when launching a new product.